% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifactor.R
\name{bifactor}
\alias{bifactor}
\title{Generate a Bifactor Simulation Environment}
\usage{
bifactor(...)
}
\arguments{
\item{...}{objects inherited from parent}
}
\value{
an environment stored to a list object of the bifactor simulation environment
}
\description{
Generate a Bifactor Simulation Environment
}
