% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countRhat.R
\name{countRhat}
\alias{countRhat}
\title{Rhat Convergence Indicator Function}
\usage{
countRhat(modsum, rHatThreshold = 1.05)
}
\arguments{
\item{rHatThreshold}{maximum tolerance for indicated convergence based on Rhat values}

\item{modum}{`data.frame` object generated from `$summary()` method on a `cmdstanr` model environment}
}
\value{
count of Rhat > threshold
}
\description{
Rhat Convergence Indicator Function
}
