% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDims.R
\name{getDims}
\alias{getDims}
\title{Find Dimensions of Filtered \code{.GlobalEnv} Object}
\usage{
getDims(name, envir)
}
\arguments{
\item{name}{name of target object}

\item{envir}{name of target environment}
}
\value{
integer of object's total dimensions
}
\description{
Find Dimensions of Filtered \code{.GlobalEnv} Object
}
