% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStdSumScore.R
\name{getStdSumScore}
\alias{getStdSumScore}
\title{Calculate Standardized Sum Scores}
\usage{
getStdSumScore(resps)
}
\arguments{
\item{resps}{matrix of dichotomized (0/1) item response data}
}
\value{
a vector of standardized sum scores of the measured latent trait
}
\description{
Calculate Standardized Sum Scores
}
