% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNeg.R
\name{makeNeg}
\alias{makeNeg}
\title{Negative Lambda Indicator Function}
\usage{
makeNeg(lambda, numNeg = 2)
}
\arguments{
\item{lambda}{inputted item discrimination/slope values}

\item{numNeg}{integer indicating quantity of lambda values to negate}
}
\value{
a vector of all lambda values (including negated lambdas)
}
\description{
Negative Lambda Indicator Function
}
