% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSelect.R
\name{methodSelect}
\alias{methodSelect}
\title{Method Selector by Modulo}
\usage{
methodSelect(base10, methodsMatrix)
}
\arguments{
\item{base10}{number in base-10 (decimal) representation}

\item{methodsMatrix}{matrix containing all combinations of tested methods conditions}
}
\value{
A selected row (after converting from base-10/decimal representation) of the methods matrix that describes the tested conditions
}
\description{
Method Selector by Modulo
}
