% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopl.R
\name{twopl}
\alias{twopl}
\title{Generate a 2-Parameter Logistic (2PL) IRT Simulation Environment}
\usage{
twopl(...)
}
\arguments{
\item{...}{objects inherited from parent}
}
\value{
an environment stored to a list object of the bifactor simulation environment
}
\description{
Generate a 2-Parameter Logistic (2PL) IRT Simulation Environment
}
